chore: moved Service\Server namespace in Security\OAUth2 chore: adapted documentation
